Require Postgres 10 and support automatic updates to newer Postgres versions
If the Postgres base image is updated (like in this commit to version
10), the entry point script will detect this and exist with a warning.
This warning makes it clear that if the `DB_UPDATE=true` environment
variable is set in the `docker-compose.yml` file, the database update
happens automatically. It is advisable to make a backup before this
automatic update is performed, because a hard-link based upgrade is
performed to not require more space. The documentation has more
information on this.

See catmaid/CATMAID#1723