Make sure postgres superuser exists also in regular runs
Having a postgres superuser available is common and expected. Due to
setting POSTGRES_USER in the docker-compose.yaml file, the database is
initialized with this user as a superuser (catmaid). This is not a
problem at all, but maintenance tasks are easier to write if we can
assume a postgres user to be present. If this isn't the case, this
superuser will now be created on startup.